requirements.txt
docxtpl
python-docx
jinja2
python-dotenv